requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
jax==0.4.26
kiwisolver==1.4.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
multidict==6.0.5
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
paho-mqtt==2.0.0
pillow==10.3.0
protobuf==3.20.3
pycparser==2.22
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
requests==2.31.0
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
